refactor(desktop,mobile): rename state-manager/configureStore files
