+++ .changeset/consistent-configure-store-files.md
@@ -1,0 +1,6 @@
+---
+"ledger-live-desktop": patch
+"live-mobile": patch
+---
+
+Rename store files to consistent state-manager/configureStore convention
