+++ apps/ledger-live-desktop/src/mvvm/hooks/redux.ts
@@ -4,7 +4,7 @@ import {
   useStore as useStoreBase,
 } from "react-redux";
 import type { State } from "~/renderer/reducers";
-import type { AppDispatch, ReduxStore } from "~/renderer/createStore";
+import type { AppDispatch, ReduxStore } from "~/state-manager/configureStore";
 
 export const useDispatch = useDispatchBase.withTypes<AppDispatch>();
 export const useSelector = useSelectorBase.withTypes<State>();
